Bug fixes, refactoring and configuration
The communication string with the database and its references in the repository files has been adjusted, correcting the problem of connecting the API to the server in some cases. The bug of creating tables on the first run has been fixed, the function that creates the schema has been moved to the use case that updates the database. New projects have been added to be saved. The formatting of the logger texts has been adjusted.